Script_code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
brotlipy=0.7.0
ca-certificates=2020.7.22=0
certifi=2020.6.20=py37_0
cffi=1.14.3
chardet=3.0.4
click=7.1.2
cryptography=3.1.1
gitdb=4.0.5=py_0
gitpython=3.1.9=py_0
idna=2.10=pyh9f0ad1d_0
libcxx=10.0.0=1
libedit=3.1.20191231
libffi=3.3
libprotobuf=3.6.0
mysql-connector-c=6.1.11
mysql-connector-python=8.0.18
ncurses=6.2
openssl=1.1.1h
pip=20.2.3=py37_0
protobuf=3.6.0
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=19.1.0=py_1
pysocks=1.7.1
python=3.7.7
python_abi=3.7=1_cp37m
readline=8.0
requests=2.24.0=pyh9f0ad1d_0
setuptools=50.3.0
six=1.15.0=pyh9f0ad1d_0
smmap=3.0.4=pyh9f0ad1d_0
sqlite=3.33.0
tableauserverclient=0.13
tk=8.6.10
urllib3=1.25.10=py_0
wheel=0.35.1=py_0
xz=5.2.5
zlib=1.2.11
